Release3

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -50,9 +50,10 @@ public class ImageController {
 	// Also now you need to add the tags of an image in the Model type object
 	// Here a list of tags is added in the Model type object
 	// this list is then sent to 'images/image.html' file and the tags are displayed
-	@RequestMapping("/images/{title}")
-	public String showImage(@PathVariable("title") String title, Model model) {
-		Image image = imageService.getImageByTitle(title);
+	@RequestMapping("/images/{imageId}/{title}")
+	public String showImage(@PathVariable("imageId") Integer imageId, @PathVariable("title") String title, Model model) {
+		// Image image = imageService.getImageByTitle(title);
+		Image image = imageService.getImageByID(imageId);
 		model.addAttribute("image", image);
 		model.addAttribute("tags", image.getTags());
 		return "images/image";
@@ -114,13 +115,21 @@ public class ImageController {
 	// This string is then displayed by 'edit.html' file as previous tags of an
 	// image
 	@RequestMapping(value = "/editImage")
-	public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+	public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
 		Image image = imageService.getImage(imageId);
+		if(isOperationAllowedOnImage(image,session)){
+			String tags = convertTagsToString(image.getTags());
+			model.addAttribute("image", image);
+			model.addAttribute("tags", tags);
+			return "images/edit";
+		} else {
+			String error = "Only the owner of the image can edit the image";
+			model.addAttribute("editError", error);
+			model.addAttribute("image", image);
+			model.addAttribute("tags", image.getTags());
+			return "images/image";
+		}
 
-		String tags = convertTagsToString(image.getTags());
-		model.addAttribute("image", image);
-		model.addAttribute("tags", tags);
-		return "images/edit";
 	}
 
 	// This controller method is called when the request pattern is of type
@@ -162,7 +171,7 @@ public class ImageController {
 		updatedImage.setDate(new Date());
 
 		imageService.updateImage(updatedImage);
-		return "redirect:/images/" + updatedImage.getTitle();
+		return "redirect:/images/"+updatedImage.getId()+"/" + updatedImage.getTitle();
 	}
 
 	// This controller method is called when the request pattern is of type
@@ -171,9 +180,18 @@ public class ImageController {
 	// id of the image to be deleted
 	// Looks for a controller method with request mapping of type '/images'
 	@RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-	public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-		imageService.deleteImage(imageId);
-		return "redirect:/images";
+	public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+		Image image = imageService.getImage(imageId);
+		if(isOperationAllowedOnImage(image,session)){
+			imageService.deleteImage(imageId);
+			return "redirect:/images";
+		} else {
+			String error = "Only the owner of the image can delete the image";
+			model.addAttribute("deleteError", error);
+			model.addAttribute("image", image);
+			model.addAttribute("tags", image.getTags());
+			return "images/image";
+		}
 	}
 
 	// This method converts the image to Base64 format
@@ -220,5 +238,13 @@ public class ImageController {
 		}
 
 		return tagString.toString();
+	}
+	private boolean isOperationAllowedOnImage(Image image, HttpSession session){
+		User user = (User) session.getAttribute("loggeduser");
+		
+		if(image.getUser().getId()==user.getId())// Logged in user is the owner of the image
+			return true;
+		else // Logged in user is not the owner of the image
+			return false;
 	}
 }
